Rename ArraySchema to IterableSchema
